minimize codelength of main.js by splitting into 3 files
